Zero packet loss testing has been added.
Using l2 testing, where no ARP messages are being used, you can now run
a zero packet loss test. The binary search will continue to find the
speed where absolutely no packets have been lost during the test.
On top of that, the output has been reformatted to also show the
absolute and relative packet loss.
The *.test files have now 2 new parameters: drop_rate_treshold and accuracy.
When drop_rate_treshold is set to 0, this test will conclude successfull
when no packet has been lost whatsoever. If this is set to any other
number, the test will be successfull when the relative packet loss is
lower than drop_rate_treshold.
The new accuray parameter defines when the binary search will stop:
when the relative difference between the min and max of the search
window is lower than accuracy.
The irq testing is slightly modified to print the total number of
interrupts in a certain time bucket, not only as the number per second,
but when there is no interrupt whatsowever, we print '0' in stead of
'0.0'

Change-Id: I8fafded0a225cbe1ba7ac61cd28e94d72099e4a0
Signed-off-by: Luc Provoost <luc.provoost@intel.com>